fix(daemon): deliver orchestrator .env to agent workspace; guarantee migration advisory unlock
The daemon previously cloned the repo, wrote `.env` into that clone, and then
`runInlinePipeline()` cloned a second time and ran the agent there — so every
`envVars` map was dropped on the floor, and the `.cred.sh` / `.env` in the
outer throwaway dir was the only target of cancel/exit cleanup.

Collapse to a single clone: thread `envVars` through `BotContext`, let the
pipeline write `.env` after its own checkout, and expose the live `workDir`
to the daemon via a new `onWorkDirReady` override so cancel and SIGKILL paths
still have something to rm. Matches CodeRabbit PR #14 feedback on
`src/daemon/job-executor.ts:285`.

Also move the Postgres migration advisory unlock into a `finally` guarded by
a `locked` bool — a throwing migration previously skipped the unlock and left
the pooled connection session-locked, blocking the next migrator. Addresses
the outside-diff finding on `src/db/migrate.ts:50-85`.
